Restructure code and add new dom module script
### Added
- New modules: `dom`, `svg`, `media-queries`, and `custom-elements`

### Changed
- esQuery module now exports `{ $ }` as well as default class
- Update `functions.js` to re-export functions from new modules which it used to contain
