+++ test/index.html
@@ -47,9 +47,14 @@
 		<noscript>You obviously need JavaScript enabled to test a JavaScript library.</noscript>
 		<button type="button" is="share-button" text="The CDN for KernValley.US" url="https://cdn.kernvalley.us" class="btn btn-primary round fixed bottom right z-4" hidden="">
 			<svg width="16" height="16" fill="currentColor" class="icon" viewBox="0 0 16 16">
-				<path d="M5.969 7.969a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0zm9.969 5a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0zm0-10a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0z" overflow="visible"/>
-				<path d="M12.625 2.156L2.562 7.031.75 7.938l1.812.906 10.032 5.062.906-1.812-8.22-4.156 8.219-4-.875-1.782z" overflow="visible"/>
+				<use xlink:href="#public-share"></use>
 			</svg>
 		</button>
+		<svg width="0" height="0">
+			<symbol id="public-share"  viewBox="0 0 16 16">
+				<path d="M5.969 7.969a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0zm9.969 5a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0zm0-10a2.969 2.969 0 1 1-5.938 0 2.969 2.969 0 1 1 5.938 0z" overflow="visible"/>
+				<path d="M12.625 2.156L2.562 7.031.75 7.938l1.812.906 10.032 5.062.906-1.812-8.22-4.156 8.219-4-.875-1.782z" overflow="visible"/>
+			</symbol>
+		</svg>
 	</body>
 </html>
